requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-allauth==0.51.0
django-environ==0.9.0
django-filter==22.1
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
kombu==5.2.4
MarkupSafe==2.1.1
model-bakery==1.7.0
oauthlib==3.2.1
openapi-codec==1.3.2
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-cryptography==3.3.23
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
